Reduce load caused by health endpoint (#1124)
